refactor(consumer): reconnect timeout, service structure
