+++ example.env
@@ -8,3 +8,5 @@ API_PORT=8000
 
 WS_HOST=websocket
 WS_PORT=8001
+
+CONSUMER_RECONNECT_TIMEOUT=5
