+++ internal/logging.py
@@ -1,0 +1,8 @@
+import logging
+
+logging.basicConfig(
+    format='%(asctime)s: %(levelname)s - %(message)s',
+    level=logging.INFO
+)
+
+logger = logging.getLogger()
